% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPCQ_function.R
\name{func_iPCQ}
\alias{func_iPCQ}
\title{All costprices following the costing manual}
\usage{
func_iPCQ(dat, recall_weeks, reference_year, yeardays = 365.25)
}
\description{
All costprices following the costing manual
}
