% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tic-P_function.R
\name{func_ticP}
\alias{func_ticP}
\title{All costprices following the costing manual excluding medication'}
\usage{
func_ticP(dat, reference_year)
}
\description{
All costprices following the costing manual excluding medication'
}
