% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ipcq_app.R
\name{run_ipcq_app}
\alias{run_ipcq_app}
\title{Run the iPCQ Shiny calculator}
\usage{
run_ipcq_app(launch.browser = TRUE, ...)
}
\arguments{
\item{launch.browser}{Passed to \code{shiny::runApp()}.}
\item{...}{Additional arguments passed to \code{shiny::runApp()}.}
}
\value{The value returned invisibly by \code{shiny::runApp()}.}
\description{Launches the Shiny application included with TforT.}
